Make render and opengl agree on camera left, right, top, bottom
(fix integer truncation issue in calculations.)

+++ entity.c
@@ -924,7 +924,7 @@ void camera_set_parameters(struct entity_context *cx, float near, float far,
 
 	/* based on gluPerspective to find the right, left, top, and bottom */
 	float scale = tan(cx->camera.angle_of_view * 0.5 * M_PI / 180.0) * cx->camera.near;
-	cx->camera.right = xvpixels/yvpixels * scale;
+	cx->camera.right = ((float) xvpixels / (float) yvpixels * scale);
 	cx->camera.left = -cx->camera.right;
 	cx->camera.top = scale;
 	cx->camera.bottom = -cx->camera.top;
